Improve output path resolution during build (#13222)
The build.sh script for msquic assumes an output path with a
configurable name, set during the build. If that name changes, then the
build breaks when the fuzzer attempts linkage, as the libmsquic.a
library path is hard coded.

instead, pull the name from the cmake configuration and just use that,
so we don't have to hard code it here